pyproject.toml
[project]
name = "main"
version = "0.1.0"
description = ""
authors = [{ name = "bulldra", email = "bulldra@gmail.com" }]
readme = "README.md"
license = "MIT"
requires-python = ">=3.11,<4.0"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "functions-framework>=3.4.0",
    "google-cloud-logging>=3.11.3",
    "google-cloud>=0.34.0",
    "requests>=2.31.0",
    "slack-sdk>=3.27.1",
    "types-requests>=2.31.0",
    "types-beautifulsoup4>=4.12.0",
    "pytest>=8.2.0",
    "feedparser>=6.0.11",
    "pypdf>=4.2.0",
    "pydub>=0.25.1",
    "pydantic>=2.10.0",
    "openai>=1.78.0",
    "openai-agents>=0.0.11",
    "google-api-python-client>=2.167.0",
    "google-cloud-storage>=3.1.0",
    "google-genai>=1.14.0",
    "tweepy>=4.16.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.mypy]
mypy_path = "src"
explicit_package_bases = true
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest>=8.2.0",
    "black==24.4.2",
]

[tool.black]
line-length = 88
target-version = ["py312"]
